### 2019-06-25 #1521 @brmather
@kyleniemeyer thank you for handling the submission. I have removed @rdelhaye ORCID and the PDF has been successfully generated.

I recommend the following two reviewers for our submission:

- Santiago Soler (github:santisoler)
- Claire Bouligand Claire.Bouligand@univ-grenoble-alpes.fr

Please let me know if you would like more suggestions.

Kind regards,
Ben
