### 2019-06-25 #1521 @lheagy
:wave: @fourndo, @santisoler, would you be willing to review this submission for JOSS (Journal of Open Source Software)? With both of your experience in Magnetic methods, I would very much appreciate your input!

If you haven't reviewed with JOSS before, the review process is a bit different. We focus on the software, and provide a checklist to help guide the review (more info in the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)).

If you don't have the time at the moment, I would appreciate any recommendations you have. Thanks!
