### 2019-06-27 #1521 @santisoler
Hi @lheagy and @brmather! I'll be glad to review `pycurious`.
I'll try to start opening issues for the end of next week. Please be patient while I read through the submitted paper and explore the repository.
