### 2019-06-27 #1521 @lheagy
Just a quick update on a second reviewer: I have sent an email to Claire and am waiting to hear back. I will keep you posted. @santisoler: if you would like to get going on the checklist right away, please let me know and I can start the review. Otherwise I will hopefully find a second reviewer and we can start the process then.
