### 2019-06-27 #1521 @santisoler
@lheagy I'm not planning to get into the review until the end of next week, so there's no rush to start it. But if you want to make my checklist available right away so you don't have to remember to do it in the future, go ahead.
