### 2019-07-01 21f006f
Slight improvements to documentation
