### 2019-07-01 380728f
Slight improvements to documentation
