### 2019-07-01 5726f15
Slight improvements to documentation
