### 2019-07-02 #1544 @whedon
[REVIEW]: PyCurious: A Python module for computing the Curie depth from the magnetic anomaly.
**Submitting author:** @brmather (<a href="http://orcid.org/0000-0003-3566-1557">Ben Mather</a>)
**Repository:** <a href="https://github.com/brmather/pycurious" target ="_blank">https://github.com/brmather/pycurious</a>
**Version:** 1.0.2
