### 2019-07-02 #1521 @lheagy
:wave: @jessepisel, would you be willing to review this submission for JOSS?

If you haven't reviewed with JOSS before, the review process is a bit different. We focus on the software, and provide a checklist to help guide the review (more info in the reviewer guidelines).

If you don't have the time at the moment, I would appreciate any recommendations you have. Thanks!
