### 2019-07-03 #1521 @jessepisel
Howdy @lheagy I would be happy to review this submission for JOSS. I haven't reviewed with JOSS yet so I will go through the reviewer guidelines and checklist and get started in the next couple of days.
