### 2019-07-03 #1521 @lheagy
Excellent, many thanks @jessepisel! The review issue is started over in #1544. I will add a checklist for you there. In order to be able to check off items, please accept the invitation to JOSS: https://github.com/openjournals/joss-reviews/invitations. Note that you will likely want to "unwatch" the repo so that you don't get swamped with all of the review activity!
