### 2019-07-08 d03795a
JOSS edits
I had a few minor edits for the JOSS review. The first edit was a typo in paper.md. The second was in the power spectrum example 1. The third was reformatting the docstring for `radial_spectrum_log`. Other than those it looks pretty good!
