### 2019-07-08 #1544 @lheagy
Thanks for your review @jessepisel! I see pr brmather/pycurious#12 is active and am linking it here so we can keep track of the progress
