### 2019-07-09 #1544 @brmather
Thanks for your review @jessepisel - I have merged your pull request into the master branch and addressed your comments in brmather/pycurious#12. If you notice anything else that requires attention, please feel free to open an issue ticket.
