### 2019-07-10 #1544 @santisoler
Hi @brmather! Sorry for the delay. I'll start opening some issues on `pycurious`' repository. We can discuss them over there.
