### 2019-07-10 brmather/pycurious#13 @santisoler
[REVIEW] Cython building fails
Part of https://github.com/openjournals/joss-reviews/issues/1544.

When trying to install `pycurious` with pip on Python 3 (Anaconda) I get the following error:

```
$ pip install pycurious
Collecting pycurious
  Using cached https://files.pythonhosted.org/packages/2a/f4/063241be23baf131ef7d6967d26d353fa9a00eb4fe0751db9b991df944e4/pycurious-0.4.tar.gz
Requirement already satisfied: numpy in ./.anaconda3/envs/pycurious/lib/python3.7/site-packages (from pycurious) (1.16.4)
Requirement already satisfied: scipy>=0.15.0 in ./.anaconda3/envs/pycurious/lib/python3.7/site-packages (from pycurious) (1.3.0)
Requirement already satisfied: Cython>=0.25.2 in ./.anaconda3/envs/pycurious/lib/python3.7/site-packages (from pycurious) (0.29.12)
Building wheels for collected packages: pycurious
  Building wheel for pycurious (setup.py) ... error
  ERROR: Complete output from command /home/santi/.anaconda3/envs/pycurious/bin/python -u -c 'import setuptools, tokenize;__file__='"'"'/tmp/pip-install-1e1fzccz/pycurious/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' bdist_wheel -d /tmp/pip-wheel-dyyg25o2 --python-tag cp37:
  ERROR: running bdist_wheel
  running build
  running build_py
  creating build
  creating build/lib.linux-x86_64-3.7
  creating build/lib.linux-x86_64-3.7/pycurious
  copying pycurious/download.py -> build/lib.linux-x86_64-3.7/pycurious
  copying pycurious/grid.py -> build/lib.linux-x86_64-3.7/pycurious
  copying pycurious/__init__.py -> build/lib.linux-x86_64-3.7/pycurious
  copying pycurious/documentation.py -> build/lib.linux-x86_64-3.7/pycurious
  copying pycurious/mapping.py -> build/lib.linux-x86_64-3.7/pycurious
  copying pycurious/optimise.py -> build/lib.linux-x86_64-3.7/pycurious
  creating build/lib.linux-x86_64-3.7/pycurious/Examples
  creating build/lib.linux-x86_64-3.7/pycurious/Examples/data
  copying pycurious/Examples/data/test_mag_data.txt -> build/lib.linux-x86_64-3.7/pycurious/Examples/data
  copying pycurious/Examples/0-StartHere.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples
  creating build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks
  creating build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Tanaka
  copying pycurious/Examples/Notebooks/Tanaka/Ex3-Parameter-exploration.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Tanaka
  copying pycurious/Examples/Notebooks/Tanaka/Ex2-Compute-Curie-depth.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Tanaka
  copying pycurious/Examples/Notebooks/Tanaka/Ex1-Plot-power-spectrum.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Tanaka
  creating build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Bouligand
  copying pycurious/Examples/Notebooks/Bouligand/Ex4-Spatial-variation-of-Curie-depth.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Bouligand
  copying pycurious/Examples/Notebooks/Bouligand/Ex3-Posing-the-inverse-problem.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Bouligand
  copying pycurious/Examples/Notebooks/Bouligand/Ex2-Compute-Curie-depth.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Bouligand
  copying pycurious/Examples/Notebooks/Bouligand/Ex5-Mapping-Curie-depth-EMAG2.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Bouligand
  copying pycurious/Examples/Notebooks/Bouligand/Ex1-Plot-power-spectrum.ipynb -> build/lib.linux-x86_64-3.7/pycurious/Examples/Notebooks/Bouligand
  running build_ext
  building 'pycurious.radon' extension
  creating build/temp.linux-x86_64-3.7
  creating build/temp.linux-x86_64-3.7/src
  gcc -pthread -B /home/santi/.anaconda3/envs/pycurious/compiler_compat -Wl,--sysroot=/ -Wsign-compare -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -fPIC -I/home/santi/.anaconda3/envs/pycurious/lib/python3.7/site-packages/numpy/core/include -Isrc -I/home/santi/.anaconda3/envs/pycurious/include/python3.7m -c src/radon.c -o build/temp.linux-x86_64-3.7/src/radon.o -std=c99
  In file included from /home/santi/.anaconda3/envs/pycurious/lib/python3.7/site-packages/numpy/core/include/numpy/ndarraytypes.h:1822,
                   from /home/santi/.anaconda3/envs/pycurious/lib/python3.7/site-packages/numpy/core/include/numpy/ndarrayobject.h:12,
                   from /home/santi/.anaconda3/envs/pycurious/lib/python3.7/site-packages/numpy/core/include/numpy/arrayobject.h:4,
                   from src/radon.c:625:
  /home/santi/.anaconda3/envs/pycurious/lib/python3.7/site-packages/numpy/core/include/numpy/npy_1_7_deprecated_api.h:17:2: warning: #warning "Using deprecated NumPy API, disable it with " "#define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION" [-Wcpp]
     17 | #warning "Using deprecated NumPy API, disable it with " \
        |  ^~~~~~~
  gcc -pthread -B /home/santi/.anaconda3/envs/pycurious/compiler_compat -Wl,--sysroot=/ -Wsign-compare -DNDEBUG -g -fwrapv -O3 -Wall -Wstrict-prototypes -fPIC -I/home/santi/.anaconda3/envs/pycurious/lib/python3.7/site-packages/numpy/core/include -Isrc -I/home/santi/.anaconda3/envs/pycurious/include/python3.7m -c src/cradon.c -o build/temp.linux-x86_64-3.7/src/cradon.o -std=c99
  src/cradon.c: In function ‘radon2d’:
  src/cradon.c:228:9: warning: returning ‘void *’ from a function with return type ‘int’ makes integer from pointer without a cast [-Wint-conversion]
    228 |         import_array();  // to use PyArray_SimpleNewFromData
        |         ^~~~~~~~~~~~
  gcc -pthread -shared -B /home/santi/.anaconda3/envs/pycurious/compiler_compat -L/home/santi/.anaconda3/envs/pycurious/lib -Wl,-rpath=/home/santi/.anaconda3/envs/pycurious/lib -Wl,--no-as-needed -Wl,--sysroot=/ build/temp.linux-x86_64-3.7/src/radon.o build/temp.linux-x86_64-3.7/src/cradon.o -Lpycurious -o build/lib.linux-x86_64-3.7/pycurious/radon.cpython-37m-x86_64-linux-gnu.so
  /home/santi/.anaconda3/envs/pycurious/compiler_compat/ld: build/temp.linux-x86_64-3.7/src/radon.o: unable to initialize decompress status for section .debug_info
  /home/santi/.anaconda3/envs/pycurious/compiler_compat/ld: build/temp.linux-x86_64-3.7/src/radon.o: unable to initialize decompress status for section .debug_info
  /home/santi/.anaconda3/envs/pycurious/compiler_compat/ld: build/temp.linux-x86_64-3.7/src/radon.o: unable to initialize decompress status for section .debug_info
  /home/santi/.anaconda3/envs/pycurious/compiler_compat/ld: build/temp.linux-x86_64-3.7/src/radon.o: unable to initialize decompress status for section .debug_info
  build/temp.linux-x86_64-3.7/src/radon.o: file not recognized: file format not recognized
  collect2: error: ld returned 1 exit status
  error: command 'gcc' failed with exit status 1
  ERROR: Failed building wheel for pycurious
  Running setup.py clean for pycurious
Failed to build pycurious
```

I don't think the two raised warnings are causing the error. I'm more worried about the `file format not recognized` error when reading `radon.o`.

Do you know what am I doing wrong or how to solve this?
