### 2019-07-10 brmather/pycurious#15 @santisoler
[REVIEW] Specify meaning for x and y when defining the grid
Part of https://github.com/openjournals/joss-reviews/issues/1544.

On `CurieGrid` docstring, would be nice to specify what `x` and `y` stands for.
Some users may be used to packages that uses `y` for easting and `x` for northing, while others may refer to a classical Cartesian axes, where `x` is usually assigned to the horizontal (easting) coordinate and `y` to the vertical one (northing).

Also, would be nice to specify the expected shape of the `grid` 2d-array: `(ny, nx)`.
