### 2019-07-10 brmather/pycurious#16 @santisoler
[REVIEW] Raise error if window is too big
Part of https://github.com/openjournals/joss-reviews/issues/1544.

When extracting a subgrid, an error is raised if its center falls outside the data range:
https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L96-L98

That's a nice thing to check! I would add another check: if the window is too big it may reach regions beyond the data range, what may cause errors afterwards but would be hard to track where they were originated.
What do you think?
