### 2019-07-10 brmather/pycurious#17 @santisoler
[REVIEW] Unnecessary allocation of memory?
Part of https://github.com/openjournals/joss-reviews/issues/1544.

Is there a special reason why `data` is copied on the following line?
https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L180

Remember that the `copy()` method allocates memory, what would increase the memory consumption if `data` is too big and require more computation time.
