### 2019-07-10 brmather/pycurious#18 @santisoler
[REVIEW] Is scale argument necessary on radial_spectrum?
Part of https://github.com/openjournals/joss-reviews/issues/1544.

I'm a little bit confused about the `scale` argument of `CurieGrid.radial_spectrum`:
https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L190

Is it only used for units conversion? If that's the case, it's not a good choice to let users to change unit conversions constants.

By the way, is there a special reason why converting units from meters (grid coordinates) to km (wavenumbers and estimated depths)?
