### 2019-07-10 brmather/pycurious#19 @santisoler
[REVIEW] Avoid repeating code
Part of https://github.com/openjournals/joss-reviews/issues/1544.

Methods `CurieGrid.radial_spectrum`, `CurieGrid.radial_spectrum_log` and `CurieGrid.azimuthal_spectrum` perform almost the same task. They differ only on a few lines. It's a good practice to avoid repeating code. It makes maintenance and testing a lot easier.

Would you consider creating a private method for `CurieGrid` that performs common tasks for these three methods, and refactor them to just call this new private method?
