### 2019-07-10 brmather/pycurious#20 @santisoler
[REVIEW] Extend documentation for power spectrum methods
Part of https://github.com/openjournals/joss-reviews/issues/1544.

`CuriedGrid` has the following methods: `radial_spectrum` and `radial_spectrum_log`.
Their names suggest the second one computes the radial spectrum and applies the `np.log`, while the first one does not. Nevertheless, the difference between both of them does not rely on the `np.log` but on the `np.sqrt`:

**`CurieGrid.radial_spectrum`**
https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L250

**`CurieGrid.radial_spectrum_log`**
https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L318

Moreover, I don't quite understand what's the goal of `CurieGrid.azimuthal_spectrum`. That could be related to some lack of knowledge on the subject from my part.

I think extending their docstring would help to a better understanding of what they are doing. Maybe including some math and references would be nice. Renaming the methods in order to make very clear what they do it's also a good approach.

Am I understanding something wrong here or do you agree with my point of view?
