### 2019-07-10 f19e08f
Fix styling when checking if variable is None
