### 2019-07-10 702a8f3
Rename nw to nr and remove unneeded nw variable
