### 2019-07-10 825507e
Remove redundant zero on range
