### 2019-07-10 brmather/pycurious#21 @santisoler
[REVIEW] Is Zo the bottom or centroid?
Part of https://github.com/openjournals/joss-reviews/issues/1544.

Sometimes the docstrings refer to `Zor` as the *bottom of the magnetic source*.  Nevertheless, when computing `Zb`, `ComputeTanaka()` applies the following:
https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L665
It agrees with Equation 7 (Tanaka et al., 1999), at which `Z_b` is the depth to the bottom and `Z_0` is the depth to the centroid.

Is the docstring mistaken or I'm missing something here?

## References
- Tanaka A., Okubo Y., Matsubayashi O. (1999), Curie point depth based on spectrum analysis of the magnetic anomaly data in East and Southeast Asia, Tectonophysics, doi: 10.1016/S0040-1951(99)00072-4
