### 2019-07-10 brmather/pycurious#22 @santisoler
[REVIEW] Fix wrong error propagation
Part of https://github.com/openjournals/joss-reviews/issues/1544.

When computing the error for the depth to the bottom of the magnetic source, the errors of `Zor` and `Ztr` should be added instead of subtracted:
https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L666

```python
 dZb = 2.0*dZor + dZtr
```
