### 2019-07-10 a90018f
Remove unused variables and imported functions
