### 2019-07-10 brmather/pycurious#23 @santisoler
[REVIEW] Unused variables on tanaka1999()
Part of https://github.com/openjournals/joss-reviews/issues/1544.

There are two unused variables on `tanaka1999()`:

https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L607

https://github.com/brmather/pycurious/blob/759fc520c22503f0b40b6a63faa4e632dd9f1e88/pycurious/grid.py#L623

Are they not necessary or are missing on any following line?
