### 2019-07-10 #1544 @santisoler
@brmather I think `pycurious` it's a great asset for geoscientists. Curie Point Depth is a widely used tool, but there aren't well maintained and documented open source libraries to calculate it. I'm glad to review it.

I've opened some issues, although I haven't found big problems with `pycurious`. I need to finish reviewing it, but feel free to answer them and even open PRs while I'm still doing it. Sorry for my brevity, I wanted to compensate my delay, so I opened the issues very quickly.
