### 2019-07-11 brmather/pycurious#13 @brmather
Hmm, this is troubling. I am not able to replicate this issue on my machine, but after some Googling, I found this:

https://github.com/Anaconda-Platform/anaconda-project/issues/183#issuecomment-462796564

It looks like a problem with Anaconda's use of gcc, that may be remedied by installing fresh compilers:

```bash
conda install gxx_linux-64
```

Does that work for you?
