### 2019-07-11 brmather/pycurious#13 @santisoler
Thank you very much @brmather! It solved it right away! Nice finding!
Do you think these instructions could be added to the documentation?
In case you want to add instructions for installing `pycurious` on Anaconda, I can help you writing an `environment.yml` file.

Nevertheless, this issue is solved, so you can close it if you want to.
