### 2019-07-11 brmather/pycurious#13 @brmather
That's great! Glad it solved your problem. From memory another user contacted me with a similar problem, so I'll add these instructions to the documentation to avoid further headaches. If you have some advice on how to get started with the setting up an `environment.yml` file for Anaconda then that would also be very helpful. Thanks!
