### 2019-07-11 brmather/pycurious#13 @santisoler
Ok, that's a good choice. A lot of people use Anaconda. Personally I like how it manages environments, so I can quickly use any package without worrying about dependency issues.
I'll open a PR adding the `environment.yml` file and some instructions on `README.md`.
