### 2019-07-12 brmather/pycurious#13 @brmather
Thanks @santisoler for taking the time to add installation instructions for those using Anaconda. I've merged your pull request into master.
