### 2019-07-12 585fd43
Check equal node spacing
