### 2019-07-12 2a285f3
Check equal node spacing
