### 2019-07-12 27c40d5
Updatd docstring for CurieGrid
