### 2019-07-12 2896c0e
raise error if window is too big
