### 2019-07-12 brmather/pycurious#17 @brmather
Yes this was a conscious decision to ensure that `data` is not overwritten in place. The alternative is to replace

```python
data.flat[:] -= np.dot(A, c)
```
with
```python
data_new = data.flat[:] - np.dot(A,c)
return data_new.reshape(data.shape)
```
but that's no more efficient.

However, revisiting this method was good because I replaced `np.meshgrid` with the more efficient `np.mgrid` function.
