### 2019-07-12 brmather/pycurious#18 @brmather
Yes you are correct, the only point of this is to change metres into kilometres. That's because both Tanaka and Bouligand estimate Curie depth with the wavenumber in units of radians per km. Mainly I just keep that scale parameter in there to flag to users that they should carefully consider the units they work in. However, if you think this argument can be removed in place of more explicit instructions in the docstring, then I'm happy to do that.
