### 2019-07-12 2ea5bab
Refactoring code for computing power spectrum
Streamlines the computation of:

- `radial_spectrum`
- `radial_spectrum_log`
- `azimuthal_spectrum`
