### 2019-07-12 cff7690
Refactoring code for computing power spectrum
Streamlines the computation of:

- `radial_spectrum`
- `radial_spectrum_log`
- `azimuthal_spectrum`
