### 2019-07-12 brmather/pycurious#14 @brmather
Yes this makes sense. I haven't encountered such an issue, but that's not to say it won't happen to someone else. This has been addressed in 585fd432e1ad517b09989b87d461e59693cd5320 thanks!
