### 2019-07-12 brmather/pycurious#16 @brmather
This makes sense. I have also received suggestions to pad the subgrid with zeros if the window size exceeds the grid extent, but I think this will result in quite significant errors. Therefore I am happy to implement this check in 2896c0e301237317f3b5ba8952b6f990a3b10df3. I am closing this issue now, but feel free to reopen it if you have more queries.
