### 2019-07-12 brmather/pycurious#19 @brmather
I've implemented two private methods for `CurieGrid` in cff769036d73905629a448812853614afda611b3:

1. `_taper_spectrum` prepares the taper array and bins for computing the power spectrum
2. `_FFT_spectrum` computes the radial power spectrum using a multiplication constant that is the only different between `radial_spectrum` and `radial_spectrum_log` (see #20 for more discussion).

I'm going to close this issue, but lets continue this discussion in #20
