### 2019-07-12 brmather/pycurious#15 @brmather
Yes, I think it should be made as clear as possible that eastings and northings should be used (not longitude / latitude!) I have updated the docstring for `CurieGrid` in 27c40d5cd2685123c5d416d860220b248a86ba5d to elaborate that `x` and `y` are eastings and northings in metres. Also I've listed all the attributes available under the `CurieGrid` object. Thanks!
