### 2019-07-12 brmather/pycurious#20 @brmather
Thank you for raising this issue @santisoler - it is good to revisit these `CurieGrid` methods with fresh :eyes: because `radial_spectrum` and `radial_spectrum_log` are, as you say, identical except for one line. The difference comes down to slight differences between the Bouligand *et al.* (2009) and Tanaka *et al.* (1999) methods to compute Curie depth - i.e. squaring the FFT of the magnetic anomaly versus taking the square root.

It is helpful to understand that
```python
2*log(FT) == log(FT**2)
```
thus I was able to refactor the code in #19 to pass a constant parameter to control whether we are taking the square or square root:

https://github.com/brmather/pycurious/blob/cff769036d73905629a448812853614afda611b3/pycurious/grid.py#L269

Setting `const=0.5` is the same as `np.log(np.sqrt(FT[mask]))`.
