### 2019-07-12 brmather/pycurious#20 @brmather
Regarding `azimuthal_spectrum`, this method is essentially used to polarise the magnetic anomaly for a range of azimuths. The azimuths are binned according to the angle increment `theta` (in degrees). So you essentially retrieve a stack of power spectra that correspond to different angles of polarisation.

@rdelhaye perhaps you would like to elaborate on this and improve the docstring for `azimuthal_spectrum` and also at the end of notebook 1. While we didn't get very far, if I remember correctly you wanted to use this to estimate Curie depth in regions that exhibit a strong magnetic orientation...? In any case we should put more maths in our docstrings.
