### 2019-07-12 53b70fa
Changed pdoc to latex_math branch
Also removed homebrew installation of python because that is
causing errors for some reason...
