+++ .travis.yml
@@ -21,13 +21,9 @@ matrix:
       osx_image: xcode10.2
       language: generic
 
-before_install:
-  - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew update ; fi
-  - if [ "$TRAVIS_OS_NAME" = "osx" ]; then brew install python3 ; fi
-
 install:
-  - python3 -m pip install Cython numpy scipy pytest git+https://github.com/lmoresi/pdoc.git
-  - python3 -m pip install -e .
+  - pip3 install Cython numpy scipy pytest git+https://github.com/kernc/pdoc.git@latex_math
+  - pip3 install -e .
 
 script:
   # run the tests
