### 2019-07-12 brmather/pycurious#17 @santisoler
There's no special need for defining `data_new`: it is just going to be returned on the next line. So I think this should work:
```python
def remove_trend_linear(self, data):
    """
    Remove the best-fitting linear trend from the data
    This may come in handy if the magnetic data has not been
    reduced to the pole.
    Args:
        data : 2D numpy array
    Returns:
     data : 2D numpy array
    """
    nr, nc = data.shape
    yq, xq = np.mgrid[0:nc,0:nr]
    A = np.c_[xq.ravel(), yq.ravel(), np.ones(xq.size)]
    c, resid, rank, sigma = np.linalg.lstsq(A, data.ravel(), rcond=None)
    return data - np.dot(A, c).reshape(data.shape)
```
