### 2019-07-12 brmather/pycurious#18 @santisoler
Ok. I would remove the `scale` argument. It shouldn't be modifiable by the user because it's a factor used on unit conversions. It's always better to let users change units outside functions instead of passing a factor to do so (that's a big source of errors).

Although I'm more inclined to standardize the use of units (I always prefer using SI units unless it's strictly necessary - e.g. use of mGal in gravity or nT in magnetism), I understand why you would prefer using km on the depths. Nevertheless, whenever you are returning not standard units, I would add a big warning on the docstring of the function, and implicit its units on the `Returns` section.
