### 2019-07-13 d595143
Clarified docstring on ComputeTanaka function in grid.py
