### 2019-07-13 52acc31
Clarified docstring on ComputeTanaka function in grid.py
