### 2019-07-13 brmather/pycurious#21 @rdelhaye
The docstring was mistaken, you are correct - Zo as input is the centroid depth of the magnetic source, whereas Zb, the bottom, is of course what we are estimating. I've corrected this in commit d595143.
