### 2019-07-13 cb8d2fd
Correction of error propagation in ComputeTanaka function in grid.py
