### 2019-07-13 brmather/pycurious#22 @rdelhaye
Completely correct, the calculation has been amended in commit cb8d2fd. Thank you!
