### 2019-07-13 377d91b
Removal of superfluous variables in tanaka1999 function in grid.py
