### 2019-07-13 brmather/pycurious#23 @rdelhaye
Thank you for noticing these - both are indeed superfluous and removed in commit 377d91b.

The Y2E2sum term I believe was mistakenly created in completeness along with the other terms for computation of the weighted line of best fit.

The S2 term is a rescaling of the power spectrum by the spatial frequency, however, at some point we changed to this rescaling in the equations for the rescaled spectrum itself.
