### 2019-07-13 brmather/pycurious#20 @rdelhaye
@brmather Thanks for looping me in on this! The intent of the radial_spectrum function, as Ben mentioned, was to examine the isotropy of the magnetic data.

One of the assumptions commonly made in literature with the Tanaka-style spectral analysis is that the magnetisation is essentially isotropic, with the spatial Fourier transforms averaged with respect to azimuth. We thought that having the ability to instead examine the transforms on a more granular basis, by subdividing the transforms into bins by azimuth, would lead to at least the opportunity to test how valid the assumption is. I'm not entirely certain if spectral analysis results in an anisotropic area are a) valid and b) useful (i.e., which value to use? The radially averaged results? Or the deeper results? I would say that the shallower result, based on the assumptions that we're computing the bottom, is probably "more wrong").
