### 2019-07-14 f3adb03
Rework remove_trend_linear
