### 2019-07-14 c273e3d
Rework remove_trend_linear
