### 2019-07-15 700a789
Improve docstrings for CurieGrid and CurieOptimise
