### 2019-07-15 b6fb432
Improve docstrings for CurieGrid and CurieOptimise
