### 2019-07-15 c2e545e
Remove scale argument from power spectrum
Implicit is the transformation from eastings/northings (in metres)
to wavenumbers in units of rad/km

*This is detailed in the docstrings of each function*
