### 2019-07-15 brmather/pycurious#17 @brmather
Thanks @santisoler I have re-implemented this function accordingly in f3adb03daab9debdc739130c5caba5bc213c4789.
