### 2019-07-15 brmather/pycurious#18 @brmather
In general I agree and try to follow SI units wherever possible, however, the ubiquity of rad/km in the case of computing Curie depth has forced my hand in this case. Otherwise the `bouligand2009` or `tanaka1999` functions can not directly accept the output from `radial_spectrum` without an intermittent step of converting units. I predict this additional layer of complexity would be a source of error (and perhaps increase the number of GitHub issues!)

I have improved the documentation to clarify the conversion from m to km and removed the `scale` argument c2e545e5c9a2420ae90d126846c08c3bbc2422f6.
