### 2019-07-15 1ae9986
Remove `CurieGrid.radial_spectrum_log`
Update documentation to clarify what power should be used:

- `power=2.0` for Bouligand _et al._ (2009) computation
- `power=0.5` for Tanaka _et al._ (1999) computation
