### 2019-07-15 brmather/pycurious#20 @brmather
I think the key point to make here is that we leave open the possibility for the user to examine anisotropy of the magnetic field. In the future we can look at implementing additional Jupyter notebooks to explore this in greater detail, but for __version 1.0__ we have all the ingredients to compute Curie depth using two different approaches.
