### 2019-07-15 brmather/pycurious#20 @brmather
> `CuriedGrid` has the following methods: `radial_spectrum` and `radial_spectrum_log`.
Their names suggest the second one computes the radial spectrum and applies the `np.log`, while the first one does not. Nevertheless, the difference between both of them does not rely on the `np.log` but on the `np.sqrt`

I have submitted pull request #26 to eliminate `radial_spectrum_log` from the `CurieGrid` class and use `radial_spectrum` with the `power` argument to toggle between Bouligand or Tanaka approaches to compute Curie depth. Docstrings have been revised alongside Jupyter notebooks. Feel free to review the changes before I merge into `master`.
