+++ paper/paper.md
@@ -11,6 +11,7 @@ authors:
     orcid: 0000-0003-3566-1557
     affiliation: "1,2" # (Multiple affiliations must be quoted)
   - name: Robert Delhaye
+    orcid: 0000-0003-2128-4295
     affiliation: "2"
 affiliations:
  - name: School of Geoscience, The University of Sydney
