### 2019-07-15 brmather/pycurious#27 @santisoler
[REVIEW] Extend tests for core functionalities
Part of https://github.com/openjournals/joss-reviews/issues/1544.

Because `pycurious` relies on processing the magnetic data in order to compute the CDP:
- read/load magnetic grid,
- select a subgrid,
- compute FFT and its radial average,
- apply tappering and
- estimate depths to magnetic sources,

I consider relevant to extend tests for the core computations, just to make sure they work as expected and to prevent breaking anything on future changes.
The current test functions only test if the methods don't fail when run. That's important, but I think is also mandatory to check if they are working as expected.
I don't intent you to achieve a 100% coverage, but just add tests that assert if the result is close to the expected one.

Besides individual tests for each method, would be nice to add a synthetic magnetic data (with known depths of the magnetic source) and check if Tanaka and Bouligand methods recover the original depths.

What do you think? Do you agree?
