### 2019-07-16 12a4cd3
Fix #29
Adjust tolerance to within 1m
