### 2019-07-17 0c0bca6
Improved testing suite using pytest
