### 2019-07-17 brmather/pycurious#27 @brmather
Thanks for bringing this to my attention. In 0c0bca645757248c8c6727582b422b72916e7060 I've populated the test suite with core functionality tests to make sure these routine are working as expected. These tests read in a synthetic magnetic anomaly generated with a known fractal parameter and CPD. To summise:

- `test_subgrid` tests if a subgrid can be extracted from the total extent. The test passes if the subgrid is smaller than `mesh.data`.
- `test_FFT` tests if the FFT produces a radial power spectrum of a reasonable shape. Expected behaviour is that the spectrum must decrease with higher wavenumber (uncertainties should also decrease with wavenumber).
- `test_taper_functions` compares the shape of the radial power spectrum using a `np.hanning` and `np.hamming` window to no taper function. The tapered spectrum should have a lower average gradient.
- `test_Tanaka` tests whether a sensible solution of CPD can be extracted from a magnetic grid of known CPD using the Tanaka method.
- `test_optimisation` tests whether a sensible solution of CPD can be inverted from a magnetic grid of known CPD and fractal parameter, using the Bouligand method.
- `test_priors` compares the optimisation with and without *a priori* information. The test passes if the optimisation returns values closer to the _a priori_ distributions.

Please let me know if you can think of any improvements to these tests, or if you can think of any additional ones. I prefer not to test the `pycurious.mapping` module since its dependencies are optional.
