### 2019-07-17 brmather/pycurious#27 @santisoler
That's great @brmather! I like it very much now. I would incorporate some best practices for writing test functions.
For example, there's no need to add the `if else` statements. You could just use `assert` or `numpy.testing.assert_allclose` if you want to compare numerical results.

For example, on `test_subgrid()`:
https://github.com/brmather/pycurious/blob/0c0bca645757248c8c6727582b422b72916e7060/tests/test_grid.py#L25-L34

I would change it to:
```python
def test_subgrid():
    grid = pycurious.CurieGrid(d, xmin, xmax, ymin, ymax)

    window_size = 100e3
    subgrid = grid.subgrid(window_size, xc, yc)

    assert subgrid.shape[0] < grid.data.shape[0]
    assert subgrid.shape[1] < grid.data.shape[1]
```

Or `test_optimisation()`:
https://github.com/brmather/pycurious/blob/0c0bca645757248c8c6727582b422b72916e7060/tests/test_optimise.py#L29-L49

It may be changed to:
```python
import numpy.testing as npt

def test_optimisation():
    grid = pycurious.CurieOptimise(d, xmin, xmax, ymin, ymax)
    beta, zt, dz, C = grid.optimise(max_window, xc, yc, taper=np.hanning)

    x_opt = np.array([beta, zt, dz])

    # hard-coded parameters used to generate the magnetic anomaly
    zt0 = 0.305
    dz0 = 10.0 + zt0
    beta0 = 3.0

    x0 = np.array([beta0, zt0, dz0])

    # compare if they are close or not
    # some parameters should be more similar than others
    tol = np.array([0.3, 0.1, 2.0])

    for i in range(x0.size):
        npt.assert_allclose(x_opt[i], x[i], atol=tol[i])
```

Besides, all the global variables defined at the beginning of the files can be grouped inside a [pytest fixture](https://docs.pytest.org/en/latest/reference.html?highlight=fixture#fixtures). That way you prevent to accidentally modify them (and possible headaches in the future). Because you would need more than one element (`x`, `y`, `d`, etc) you could return a dictionary containing these arrays.

Let me know if you need any help with these!
