### 2019-07-17 brmather/pycurious#27 @santisoler
By the way, are `xc` and `yc` the coordinates of the center of the subgrid?
https://github.com/brmather/pycurious/blob/0c0bca645757248c8c6727582b422b72916e7060/tests/test_optimise.py#L22-L23

If that's the case, shouldn't those minus signs be plus signs?
