### 2019-07-18 bb6c1c8
Include `pytest.fixture` within testing suite
Refactor several tests to enrich `assert` statements
