### 2019-07-18 09ae019
Include `pytest.fixture` within testing suite
Refactor several tests to enrich `assert` statements
