+++ tests/test_grid.py
@@ -3,37 +3,31 @@ import pytest
 import pycurious
 import numpy as np
 
-# load magnetic anomaly - i.e. random fractal noise
-try:
-    mag_data = np.loadtxt("test_mag_data.txt")
-except:
-    mag_data = np.loadtxt("tests/test_mag_data.txt")
-
-nx, ny = 305, 305
-
-x = mag_data[:,0]
-y = mag_data[:,1]
-d = mag_data[:,2].reshape(ny,nx)
-
-xmin, xmax = x.min(), x.max()
-ymin, ymax = y.min(), y.max()
-
-xc = (xmax - xmin)/2
-yc = (ymax - ymin)/2
+from conftest import load_magnetic_anomaly
 
 
-def test_subgrid():
+def test_subgrid(load_magnetic_anomaly):
+    d  = load_magnetic_anomaly['mag_data']
+    xc = load_magnetic_anomaly['xc']
+    yc = load_magnetic_anomaly['yc']
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
+
     grid = pycurious.CurieGrid(d, xmin, xmax, ymin, ymax)
 
     window_size = 100e3
     subgrid = grid.subgrid(window_size, xc, yc)
 
-    if subgrid.shape[0] < grid.data.shape[0] and subgrid.shape[1] < grid.data.shape[1]:
-        print("PASSED! Subgrid successfully extracted from domain")
-    else:
-        assert False, "FAILED! Subgrid is of shape {} and domain is of shape {}".format(subgrid.shape, grid.data.shape)
+    error_msg = "FAILED! Subgrid is of shape {} and domain is of shape {}".format(subgrid.shape, grid.data.shape)
+    assert subgrid.shape[0] < grid.data.shape[0], error_msg
+    assert subgrid.shape[1] < grid.data.shape[1], error_msg
 
-def test_FFT():
+
+def test_FFT(load_magnetic_anomaly):
+    d  = load_magnetic_anomaly['mag_data']
+    xc = load_magnetic_anomaly['xc']
+    yc = load_magnetic_anomaly['yc']
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
+
     grid = pycurious.CurieGrid(d, xmin, xmax, ymin, ymax)
 
     # Take Fourier transform
@@ -51,13 +45,17 @@ def test_FFT():
     sigma_Phi2 = sigma_Phi[i3:2*i3]
     sigma_Phi3 = sigma_Phi[2*i3:]
 
-    if Phi1.mean() > Phi2.mean() > Phi3.mean() and \
-       sigma_Phi1.mean() > sigma_Phi2.mean() > sigma_Phi3.mean():
-        print("PASSED! Fast Fourier Transform produced a radial power spectrum")
-    else:
-        assert False, "FAILED! Fast Fourier Transform did not produce a valid power spectrum"
+    error_msg = "FAILED! Fast Fourier Transform did not produce a valid power spectrum"
+    assert Phi1.mean() > Phi2.mean() > Phi3.mean(), error_msg
+    assert sigma_Phi1.mean() > sigma_Phi2.mean() > sigma_Phi3.mean(), error_msg
 
-def test_taper_functions():
+
+def test_taper_functions(load_magnetic_anomaly):
+    d  = load_magnetic_anomaly['mag_data']
+    xc = load_magnetic_anomaly['xc']
+    yc = load_magnetic_anomaly['yc']
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
+
     grid = pycurious.CurieGrid(d, xmin, xmax, ymin, ymax)
 
[…]
