### 2019-07-18 brmather/pycurious#27 @brmather
Thanks @santisoler I appreciate those `if else` statements were not very useful. I've refactored the tests in 09ae019f43389e56db3c4e4ebada092b68114260 using a `pytest.fixture` to load in the magnetic grid. That's a much better way of doing it than loading it in at the start of each test script - `pytest` is great!

Also, the assertion statements are now enriched with information on the error. Take your `test_optimisation` for example, before the test would compare if the inverted parameters were within their tolerances all at once, but now it iterates through them individually making debugging a whole lot easier. I was also not aware of the `numpy.testing` module either. It provides a lot of useful information that is missing from the standard `assert` statement. I will use this for other tests in the future.
