### 2019-07-18 brmather/pycurious#27 @brmather
> By the way, are `xc` and `yc` the coordinates of the center of the subgrid?

You are right! It must have been late in the afternoon... :sweat_smile: It didn't impact on the tests because `xmin=0` and `ymin=0`, but it's fixed now. :+1:
