### 2019-07-18 brmather/pycurious#27 @santisoler
Great @brmather! I like it very much now! I think we can close this issue now.

Just wanted to leave you some tips for the future based on my experience and what I've learned developing Fatiando a Terra.

1. I think you are very skilled on coding and documenting the code. Writing the documentation before or while you write the code is one of the best practices, and you master it very well. A good advice is to write the code thinking on how you will test it. I know it's not easy at first, but is a good practice to take this into account while you are designing your code. Besides, it's good to write the test functions at the same time you write the code. So any new commit (or PR) will be based on three legs: code, documentation and tests.

1. I like how you linked commits on issues. It really helped me to track down changes. Nevertheless, you are missing one of the greatest tools that GitHub has by doing it: Pull Requests. At the beginning I think I felt the same way about PRs: *they are intended to contributors to share code*. Although that's true, contributors **must** open PRs when sharing code, opening PRs by repo owners it's an amazing tool:
    - They help you tracking changes (think your present self as a completely stranger for the future you, I'm sure you are not going to remember why did you added *that line* one year ago).
    - They make the developing process much transparent to others. You could ask for review to any other contributor (or even reviewer in case you submitted it to JOSS).
    - They make much easier to develop in parallel. Many contributors can still continue coding without worrying about random commits that are pushed into master.
    - They move the specific code discussion to the PR rather than having them on the issues. This helps to organize your discussions better, but also the PRs interface gives you a lot of tools: review, suggest changes, don't allow merge until PR is accepted by reviewer, etc.
    - They also offer the option to **Squash and Merge**. It squashes all commits inside the PR into a single one and then merge it into master. That way, running `git log` on master gives you a very concise history of changes, without ugly commit messages like "Experimenting with this new stuff" or "Fixed some typos".

    So, the workflow for **any** contributor (owners included) could be:
      1. Open an issue asking to solve a bug or asking for a new feature. Even though if you are going to fix it right away. It's a good practice of tracking down why you though *that bug* or *that new feature* must be tackled down.
      1. Open a PR with the new code solving that issue. Remember to link the issue on the PR message. The PR must have well written code with documentation and tests functions. If CI is configured, it must pass.
      1. Ask for review to any other contributor. Discuss modifications on the PR interface.
      1. When the PR is ready to be merged, Squash and Merge. Write a well explained commit message to summarize what the new code does.

1. A little thing regarding styling. Although Python is capable of understanding very different code styles, it's a good practice to follow common styling. One of the common Guiding Styles is [PEP8](https://www.python.org/dev/peps/pep-0008/). A few years ago I used to fix PEP8 errors by hand, although now days there's a tool that I love: [Black](https://github.com/python/black). It automatically reformat your code to meet PEP8 requirements, but also adds some own code styles that minimize the diffs and improve readability.

Hope these little tips can be useful! Take them as guidelines. I learned (and still learning) how to develop code by doing it and failing a lot, so feel free to correct me if you think something is wrong with them, or if you think there's a better way to do these things.
