### 2019-07-18 brmather/pycurious#27 @brmather
I appreciate your wisdom in these areas - thank you @santisoler! Your review has taught me a lot of good practises for writing and maintaining code collaboratively. For instance, I didn't know that writing Fix <issue #> in a pull request would automatically link and close the issue once the PR was accepted - very nifty!

There are many other extremely useful things I have learnt, but I would say the practise of opening a PR for code changes is an extremely valuable tool in a programmers' arsenal. Even if there are few developers, the checks and balances on pull request using CI are invaluable. As I am becoming more confident with git (and specifically GitHub) I'm rolling out this practise to all the other projects I am involved in.

`Black` looks like a very neat package, I'll take it for a test run soon. If you come across any other tools you think are great, please don't hesitate to reach out!
