+++ tests/test_grid.py
@@ -1,4 +1,3 @@
-
 import pytest
 import pycurious
 import numpy as np
@@ -7,26 +6,28 @@ from conftest import load_magnetic_anomaly
 
 
 def test_subgrid(load_magnetic_anomaly):
-    d  = load_magnetic_anomaly['mag_data']
-    xc = load_magnetic_anomaly['xc']
-    yc = load_magnetic_anomaly['yc']
-    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
+    d = load_magnetic_anomaly["mag_data"]
+    xc = load_magnetic_anomaly["xc"]
+    yc = load_magnetic_anomaly["yc"]
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly["extent"]
 
     grid = pycurious.CurieGrid(d, xmin, xmax, ymin, ymax)
 
     window_size = 100e3
     subgrid = grid.subgrid(window_size, xc, yc)
 
-    error_msg = "FAILED! Subgrid is of shape {} and domain is of shape {}".format(subgrid.shape, grid.data.shape)
+    error_msg = "FAILED! Subgrid is of shape {} and domain is of shape {}".format(
+        subgrid.shape, grid.data.shape
+    )
     assert subgrid.shape[0] < grid.data.shape[0], error_msg
     assert subgrid.shape[1] < grid.data.shape[1], error_msg
 
 
 def test_FFT(load_magnetic_anomaly):
-    d  = load_magnetic_anomaly['mag_data']
-    xc = load_magnetic_anomaly['xc']
-    yc = load_magnetic_anomaly['yc']
-    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
+    d = load_magnetic_anomaly["mag_data"]
+    xc = load_magnetic_anomaly["xc"]
+    yc = load_magnetic_anomaly["yc"]
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly["extent"]
 
     grid = pycurious.CurieGrid(d, xmin, xmax, ymin, ymax)
 
@@ -35,15 +36,15 @@ def test_FFT(load_magnetic_anomaly):
 
     # radial power spectrum should decrease with wavenumber
     # divide Phi into three sections
-    i3 = len(k)//3
+    i3 = len(k) // 3
     Phi1 = Phi[:i3]
-    Phi2 = Phi[i3:2*i3]
-    Phi3 = Phi[2*i3:]
+    Phi2 = Phi[i3 : 2 * i3]
+    Phi3 = Phi[2 * i3 :]
 
     # also sigma_Phi should decrease with wavenumber
     sigma_Phi1 = sigma_Phi[:i3]
-    sigma_Phi2 = sigma_Phi[i3:2*i3]
-    sigma_Phi3 = sigma_Phi[2*i3:]
+    sigma_Phi2 = sigma_Phi[i3 : 2 * i3]
+    sigma_Phi3 = sigma_Phi[2 * i3 :]
 
     error_msg = "FAILED! Fast Fourier Transform did not produce a valid power spectrum"
     assert Phi1.mean() > Phi2.mean() > Phi3.mean(), error_msg
@@ -51,10 +52,10 @@ def test_FFT(load_magnetic_anomaly):
 
 
 def test_taper_functions(load_magnetic_anomaly):
-    d  = load_magnetic_anomaly['mag_data']
-    xc = load_magnetic_anomaly['xc']
-    yc = load_magnetic_anomaly['yc']
-    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
+    d = load_magnetic_anomaly["mag_data"]
+    xc = load_magnetic_anomaly["xc"]
[…]
