+++ tests/test_optimise.py
@@ -1,4 +1,3 @@
-
 import pytest
 import pycurious
 import numpy as np
@@ -9,11 +8,11 @@ from conftest import load_magnetic_anomaly
 
 
 def test_optimisation(load_magnetic_anomaly):
-    d  = load_magnetic_anomaly['mag_data']
-    xc = load_magnetic_anomaly['xc']
-    yc = load_magnetic_anomaly['yc']
-    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
-    max_window = load_magnetic_anomaly['max_window']
+    d = load_magnetic_anomaly["mag_data"]
+    xc = load_magnetic_anomaly["xc"]
+    yc = load_magnetic_anomaly["yc"]
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly["extent"]
+    max_window = load_magnetic_anomaly["max_window"]
 
     grid = pycurious.CurieOptimise(d, xmin, xmax, ymin, ymax)
     beta, zt, dz, C = grid.optimise(max_window, xc, yc, taper=np.hanning)
@@ -35,16 +34,20 @@ def test_optimisation(load_magnetic_anomaly):
     err_msg = "FAILED! {} = {:.4f} is not within an acceptable tolerance of {}"
 
     for i in range(x0.size):
-        npt.assert_allclose(x_opt[i], x0[i], atol=tol[i], \
-            err_msg=err_msg.format(parameters[i], x_opt[i], tol[i]))
+        npt.assert_allclose(
+            x_opt[i],
+            x0[i],
+            atol=tol[i],
+            err_msg=err_msg.format(parameters[i], x_opt[i], tol[i]),
+        )
 
 
 def test_priors(load_magnetic_anomaly):
-    d  = load_magnetic_anomaly['mag_data']
-    xc = load_magnetic_anomaly['xc']
-    yc = load_magnetic_anomaly['yc']
-    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
-    max_window = load_magnetic_anomaly['max_window']
+    d = load_magnetic_anomaly["mag_data"]
+    xc = load_magnetic_anomaly["xc"]
+    yc = load_magnetic_anomaly["yc"]
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly["extent"]
+    max_window = load_magnetic_anomaly["max_window"]
 
     grid = pycurious.CurieOptimise(d, xmin, xmax, ymin, ymax)
     beta0, zt0, dz0, C0 = grid.optimise(max_window, xc, yc)
@@ -52,40 +55,126 @@ def test_priors(load_magnetic_anomaly):
     grid.add_prior(beta=(1.0, 0.1))
     beta1, zt1, dz1, C1 = grid.optimise(max_window, xc, yc)
 
-    assert abs(beta1 - 1.0) < abs(beta0 - 1.0), "FAILED! Optimisation with priors failed"
+    assert abs(beta1 - 1.0) < abs(
+        beta0 - 1.0
+    ), "FAILED! Optimisation with priors failed"
 
 
 def test_valid_numbers(load_magnetic_anomaly):
-    d  = load_magnetic_anomaly['mag_data']
-    xc = load_magnetic_anomaly['xc']
-    yc = load_magnetic_anomaly['yc']
-    xmin, xmax, ymin, ymax = load_magnetic_anomaly['extent']
-    max_window = load_magnetic_anomaly['max_window']
+    d = load_magnetic_anomaly["mag_data"]
+    xc = load_magnetic_anomaly["xc"]
+    yc = load_magnetic_anomaly["yc"]
+    xmin, xmax, ymin, ymax = load_magnetic_anomaly["extent"]
+    max_window = load_magnetic_anomaly["max_window"]
[…]
