+++ tests/test_0_imports.py
@@ -2,17 +2,24 @@ import pytest
 
 ## ===================
 
+
 def test_numpy_import():
     import numpy
+
     return
+
 
 def test_scipy_import():
     import scipy
+
     print("\t\t You have scipy version {}".format(scipy.__version__))
 
+
 def test_cython_import():
-	import Cython
-	return
+    import Cython
+
+    return
+
 
 def test_pycurious_modules():
     import pycurious

[…]
