### 2019-07-19 90098de
Reformatting to PEP8 standard
using [Black](https://github.com/python/black#installation-and-usage)
