+++ tests/Bouligand_forward.py
@@ -1,156 +1,162 @@
 ## Code to generate test magnetic data for PyCurious examples
-## Converted from MATLAB code by R Delhaye. 
-#Original code:
-#Claire Bouligand
-#03-19-2007
-#05-09-2014
-#Modified to use less memory
+## Converted from MATLAB code by R Delhaye.
+# Original code:
+# Claire Bouligand
+# 03-19-2007
+# 05-09-2014
+# Modified to use less memory
 
 import numpy as np
 import pycurious
 import time
 
-start_time=time.time()
-cm=1e-7
+start_time = time.time()
+cm = 1e-7
 
 ##########################################################
-#A VERIFIER: multiplier facteur cm par 4*pi
+# A VERIFIER: multiplier facteur cm par 4*pi
 ## TO VERIFY: multiplication factor "cm for 4*pi"
 ###########################################################
 ## Degree of fractal
-b=-3.0
+b = -3.0
 ## Dimension of pixels (km), should be multiples of the smallest
-dlx=1.0
-dly=1.0
-dlz=1.0
-dl=min([dlx,dly,dlz])
-dx=dlx/dl
-dy=dly/dl
-dz=dlz/dl
+dlx = 1.0
+dly = 1.0
+dlz = 1.0
+dl = min([dlx, dly, dlz])
+dx = dlx / dl
+dy = dly / dl
+dz = dlz / dl
 
-#nombre de pixels
+# nombre de pixels
 ## Number of pixels.
-nx=305
-ny=305
-nz=10
-nmax=int(max([nx*dx,ny*dy,nz*dz]))
+nx = 305
+ny = 305
+nz = 10
+nmax = int(max([nx * dx, ny * dy, nz * dz]))
 
-#texte="dimensions x : %d / y : %d / z : %d" % (nx*dlx,ny*dly,nz*dlz)
-#print(texte)
-#disp(texte)
+# texte="dimensions x : %d / y : %d / z : %d" % (nx*dlx,ny*dly,nz*dlz)
+# print(texte)
+# disp(texte)
 
-#Parametres de l'aimantation
+# Parametres de l'aimantation
 ## Parameters of magnetisation
-#M1=1         # aimantation moyenne = 1 A/m ## Average magnetisation
-#M2=10^(0.25) # std de l'aimantation = 10^(0.25) A/m ## Standard dev of
-#magnetisation
-M1=0
-M2=0.2
-#prof sommet zt # Depth Zt
-zt=0.305
-#prof base zb ## Depth Zb
-zb=zt+nz*dlz
+# M1=1         # aimantation moyenne = 1 A/m ## Average magnetisation
+# M2=10^(0.25) # std de l'aimantation = 10^(0.25) A/m ## Standard dev of
+# magnetisation
+M1 = 0
+M2 = 0.2
+# prof sommet zt # Depth Zt
+zt = 0.305
+# prof base zb ## Depth Zb
+zb = zt + nz * dlz
 
 ############################################################
-#GENERATION DE LA DISTRIB 3D D'AIMANTATION FRACTAL 
-#A PARTIR VOLUME DE DIM ID. (nmax,dl) SELON X, Y ET Z
+# GENERATION DE LA DISTRIB 3D D'AIMANTATION FRACTAL
+# A PARTIR VOLUME DE DIM ID. (nmax,dl) SELON X, Y ET Z
 
 if __name__ == "__main__":
 
-    M=np.zeros((nmax,nmax,nmax))
-    l=[0,(nmax-1)*dl,dl]
+    M = np.zeros((nmax, nmax, nmax))
+    l = [0, (nmax - 1) * dl, dl]
 
[…]
