+++ setup.py
@@ -26,20 +26,20 @@ def git_version():
     def _minimal_ext_cmd(cmd):
         # construct minimal environment
         env = {}
-        for k in ['SYSTEMROOT', 'PATH']:
+        for k in ["SYSTEMROOT", "PATH"]:
             v = os.environ.get(k)
             if v is not None:
                 env[k] = v
         # LANGUAGE is used on win32
-        env['LANGUAGE'] = 'C'
-        env['LANG'] = 'C'
-        env['LC_ALL'] = 'C'
-        out = subprocess.Popen(cmd, stdout = subprocess.PIPE, env=env).communicate()[0]
+        env["LANGUAGE"] = "C"
+        env["LANG"] = "C"
+        env["LC_ALL"] = "C"
+        out = subprocess.Popen(cmd, stdout=subprocess.PIPE, env=env).communicate()[0]
         return out
 
     try:
-        out = _minimal_ext_cmd(['git', 'rev-parse', '--short', 'HEAD'])
-        GIT_REVISION = out.strip().decode('ascii')
+        out = _minimal_ext_cmd(["git", "rev-parse", "--short", "HEAD"])
+        GIT_REVISION = out.strip().decode("ascii")
     except OSError:
         GIT_REVISION = "Unknown"
 
@@ -50,40 +50,49 @@ if PYPI_VERSION is None:
     PYPI_VERSION = git_version()
 
 this_directory = path.abspath(path.dirname(__file__))
-with open(path.join(this_directory, 'README.md')) as f:
+with open(path.join(this_directory, "README.md")) as f:
     long_description = f.read()
 
 
-ext = Extension(name               = 'pycurious.radon',
-                sources            = ['src/radon.pyx','src/cradon.c'],
-                include_dirs       = [np.get_include(),'src'],
-                library_dirs       = ['pycurious'],
-                extra_compile_args = ['-std=c99'])
+ext = Extension(
+    name="pycurious.radon",
+    sources=["src/radon.pyx", "src/cradon.c"],
+    include_dirs=[np.get_include(), "src"],
+    library_dirs=["pycurious"],
+    extra_compile_args=["-std=c99"],
+)
 
 if __name__ == "__main__":
-    setup(name              = 'pycurious',
-          long_description  = long_description,
-          long_description_content_type = 'text/markdown',
-          author            = "Ben Mather",
-          author_email      = "brmather1@gmail.com",
-          url               = "https://github.com/brmather/pycurious",
-          version           = PYPI_VERSION,
-          description       = "Python tool for computing the Curie depth from magnetic data",
-          ext_modules       = cythonize([ext]),
-          install_requires  = ['numpy', 'scipy>=0.15.0', 'Cython>=0.25.2'],
-          python_requires   = '>=2.7, >=3.5',
-          setup_requires    = ['pytest-runner', 'webdav'],
-          tests_require     = ['pytest', 'webdav'],
-          packages          = ['pycurious'],
-          package_data      = {'pycurious': ['Examples/data/test_mag_data.txt',
-                                             'Examples/*.ipynb',
-                                             'Examples/Notebooks/Tanaka/*.ipynb',
-                                             'Examples/Notebooks/Bouligand/*.ipynb',
-                                             'Examples/Scripts/*.py']},
-          classifiers       = ['Programming Language :: Python :: 2',
-                               'Programming Language :: Python :: 2.7',
-                               'Programming Language :: Python :: 3',
-                               'Programming Language :: Python :: 3.5',
-                               'Programming Language :: Python :: 3.6',
-                               'Programming Language :: Python :: 3.7']
-          )
+    setup(
+        name="pycurious",
+        long_description=long_description,
[…]

[…]
