### 2019-07-19 #1544 @santisoler
@lheagy I have checked off all my boxes, so I consider `pycurious` to be ready for publication.

Thanks for letting me review it, and thanks to the authors @brmather and @rdelhaye for the quick response to every issue I opened. I enjoy the open review system because it offers a great opportunity to learn from each other and help improving ourselves.
