### 2019-07-19 066d741
Update paper.md
A few grammatical updates to the paper.
