### 2019-07-19 #1544 @lheagy
Excellent, many thanks @santisoler and @jessepisel for taking time to review!! 🎉

@brmather, I have started a pull request with a few small corrections to the paper. I would also request that in the first paragraph, you give make a few more connections as to why the Curie depth is of interest when interpreting magnetic data. For example in or after the second sentence: would you mind adding some description as to why the Curie depth is used in the interpretation of magnetic data (e.g. it provides information on subsurface temperature, etc). I know you mention this, but it could be a bit more explicit for a non-domain expert. I left a few other comments in the pull request. Please let me know once you have had a chance to take a look.
