### 2019-07-20 e6113d5
Better description of the motivation for using `pycurious`
- Added a figure.png
- Updated some references
