### 2019-07-20 #1544 @brmather
Many thanks to @santisoler and @jessepisel for their extremely helpful reviews of `pycurious`. I have found JOSS' interactive review process very useful in improving software so that it is useful to as many people as possible. I have learnt a lot and will apply these practises to the other projects I'm actively working on. I thank the reviewers for the time they have spent on providing advice and submitting pull requests - it is very much appreciated!

Here is a summary of improvements to `pycurious` from the JOSS review:

- Updated contributions guide to encourage development on priority areas
- Anaconda installation support (and fixes for Cython build issues)
- Checking if windows that exceed the extent of the domain
- Correct propagation of errors with Tanaka *et al.* 1999 approach
- Refactor code to compute the radial power spectrum
- Updated Docstrings and Jupyter notebooks
- Updated test suite (using `pytest`) to test for expected functionality of core routines
- PEP8 compliant code

@lheagy thank you for coordinating the review process. I have responded to your pull request with number of additions I think broaden the appeal of the software package.
