### 2019-07-21 1986d37
Fixing citation styles
