### 2019-07-23 #1544 @lheagy
:wave: @brmather, if you are happy with the edits to the paper, please merge in the changes and generate a new release of the software. From there, please create an archive on [zenodo](https://zenodo.org/) or similar with the title and author list being the same as the paper and then post the version number and doi here. Thanks!
