### 2019-07-24 7f94ffc
Update paper.md (#31)
* Update paper.md

A few grammatical updates to the paper.

* Better description of the motivation for using `pycurious`

- Added a figure.png
- Updated some references

* Fixing grammar suggestions

* Fixing citation styles

* Whedon doesn't like the colons in each reference

* Richer image caption
