+++ paper/paper.bib
@@ -21,7 +21,7 @@ month = {June},
 number = {3-4},
 pages = {461--470},
 title = {{Curie point depth based on spectrum analysis of the magnetic anomaly data in East and Southeast Asia}},
-url = {http://linkinghub.elsevier.com/retrieve/pii/S0040195199000724},
+url = {https://dx.doi.org/10.1016/S0040-1951(99)00072-4},
 volume = {306},
 year = {1999}
 }
@@ -34,4 +34,16 @@ title = {{EMAG2: Earth Magnetic Anomaly Grid (2-arc-minute resolution) Version 3
 url = {https://dx.doi.org/10.7289/V5H70CVX},
 urldate = {2018-09-01},
 year = {2017}
+}
+
+@article{Mather2019,
+author = {Mather, Ben and Fullea, Javier},
+doi = {10.5194/se-10-839-2019},
+journal = {Solid Earth},
+number = {3},
+pages = {839--850},
+title = {{Constraining the geotherm beneath the British Isles from Bayesian inversion of Curie depth: integrated modelling of magnetic, geothermal, and seismic data}},
+url = {https://doi.org/10.5194/se-10-839-2019},
+volume = {10},
+year = {2019}
 }
