+++ paper/paper.md
@@ -24,11 +24,28 @@ bibliography: paper.bib
 
 # Summary
 
-Magnetic data is one of the most prevelant geophysical datasets available on the surface of the Earth. Curie depth is the depth at which rocks lose their magnetism. The most prevalent magnetic mineral is magnetite, which has a Curie point of 580°C, thus the Curie depth is often interpreted as the 580°C isotherm. Current methods to derive Curie depth first compute the (fast) Fourier transform over a square window of a magnetic anomaly that has been reduced to the pole. The depth and thickness of magnetic sources is estimated from the slope of the radial power spectrum.
+Multiple geophysical methods have been proposed to resolve the thermal structure of the Earth's lithosphere with varying degrees of precision.
+Geotherms may be constructed from heat flow or xenolith data, but the spatial coverage of these are often limited.
+Seismic velocity has proven effective to infer upper-mantle temperature, but its application relies on building a compositional model suitable for the geological context and estimating attenuation from grainsize and water content.
+In contrast, magnetic data are among the most widespread geophysical datasets available on the surface of the Earth and can offer useful insight into its thermal structure. 
 
-Our Python package, `pycurious`, ingests maps of the Earth's magnetic anomaly and distributes the computation of Curie depth across multiple CPUs. `pycurious` implements the [@Tanaka1999] and [@Bouligand2009] methods for computing the thickness of a buried magnetic source. The former selects portions of the radial power spectrum in the low and high frequency domain to compute the depth of magnetic sources, while the latter fits an analytical solution to the entire power spectrum. We cast the [@Bouligand2009] method within a Bayesian framwork to estimate the uncertainty of Curie depth calculations. Common computational workflows and geospatial manipulation of magnetic data are covered in the Jupyter notebooks bundled with this package. The `mapping` module includes a set of functions that help to wrangle maps of the magnetic anomaly into a useable form for `pycurious`. Such an approach is commonly encountered for transforming global compilations of the magnetic anomaly, e.g. EMAG2 [@Meyer2017], from latitudinal/longitudinal coordinates to a local projection in eastings/northings.
+The magnetic anomaly is the observed magnetic field less the contribution from the core and external fields.
+Geologic features enhance or depress the local magnetic field; this response is primarily controlled by the concentration of magnetite in the Earth's crust.
+Magnetite is the most prevalent magnetic mineral, in terms of both its magnetic susceptibility and concentration in the crust, and has a Curie point of 580°C.
+The depth at which rocks lose their permanent magnetic properties is referred to as the Curie depth.
+Due to the prevalence of magnetite in the crust, the Curie depth is often interpreted as the 580°C isotherm.
+Current methods to derive the Curie depth first compute the (fast) Fourier transform over a square window of a magnetic anomaly that has been reduced to the pole.
+The depth and thickness of magnetic sources is estimated from the slope of the radial power spectrum.
 
-<!-- `pycurious` is an object-oriented Python package that accepts grids of the Earth's magnetic anomaly and computes the Curie depth across a specified window size. The computation is distributed across all available processors to improve efficiency over large study areas.  -->
+Our Python package, `pycurious`, ingests a map of the Earth's magnetic anomaly and distributes the computation of Curie depth across multiple CPUs.
+`pycurious` implements the @Tanaka1999 and @Bouligand2009 methods for computing the thickness of a buried magnetic source.
+The former selects portions of the radial power spectrum in the low and high frequency domain to compute the depth of magnetic sources, while the latter fits an analytical solution to the entire power spectrum.
[…]
