### 2019-07-24 4286e1c
Bump Dockerfile to 1.0
