### 2019-07-24 29d88af
Update to version 1.0.2 to match Zenodo (hopefully!)
