### 2019-07-24 #1544 @brmather
Thanks @lheagy

I have released version 1.0.2. Some unknown error foiled the 1.0 release on Zenodo, so I had to bump the version to 1.0.2. The Docker image is on Docker Hub, the release is on PyPI, and the Zenodo doi is 10.5281/zenodo.3349511.
