### 2019-07-24 #1544 @lheagy
Thanks @brmather! Would you mind updating the title on the zenodo record to match the paper? e.g. "PyCurious: A Python module for computing the Curie depth from the magnetic anomaly" . Please also take a look at the proof above and give a :+1: if you are happy with it. From there, we can proceed with publishing 🎉
