### 2019-07-24 #1544 @brmather
I keep getting an internal server error on Zenodo preventing me from updating the title and author fields. Aside from that, the proof looks good.

![Screenshot_2019-07-25 Zenodo](https://user-images.githubusercontent.com/15915881/61835580-bcbcff00-aebf-11e9-9c33-2101789571ac.png)
