### 2019-07-24 #1544 @lheagy
Hmm... Is it possible you have it open in multiple tabs? (I am seeing the banner at the top that says "There us a new version of the upload being edited here")
