### 2019-07-25 #1544 @lheagy
@openjournals/joss-eics, this submission is ready to be published! Congrats @brmather 🎉
