### 2019-07-25 #1544 @kyleniemeyer
Hi @brmather, please merge the small PR I just submitted with some fixes in the paper: https://github.com/brmather/pycurious/pull/33
