### 2019-07-26 #1544 @kyleniemeyer
@brmather congrats on your article's publication in JOSS!

Thanks to  @santisoler and @jessepisel for reviewing, and to @lheagy for editing!
