### 2019-07-26 #1544 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
